models/dummy_model.py
import torch
import torch.nn as nn
import numpy as np
from PIL import Image
import torchvision.transforms as transforms

class DummyMedicalModel(nn.Module):
    def __init__(self):
        super(DummyMedicalModel, self).__init__()
        self.conv1 = nn.Conv2d(1, 16, kernel_size=3, padding=1)
        self.pool = nn.MaxPool2d(2, 2)
        self.conv2 = nn.Conv2d(16, 32, kernel_size=3, padding=1)
        self.fc1 = nn.Linear(32 * 56 * 56, 128)
        self.fc2 = nn.Linear(128, 3)  # 3 classes for demonstration
        
    def forward(self, x):
        x = self.pool(torch.relu(self.conv1(x)))
        x = self.pool(torch.relu(self.conv2(x)))
        x = x.view(-1, 32 * 56 * 56)
        x = torch.relu(self.fc1(x))
        x = self.fc2(x)
        return x

class MedicalImageAnalyzer:
    def __init__(self):
        self.model = DummyMedicalModel()
        self.transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.Grayscale(num_output_channels=1),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485], std=[0.229])
        ])
        self.classes = ['Normal', 'Pneumonia', 'COVID-19']
        
    def predict(self, image_path):
        # Load and preprocess image
        image = Image.open(image_path)
        image_tensor = self.transform(image).unsqueeze(0)
        
        # Get dummy prediction
        with torch.no_grad():
            outputs = self.model(image_tensor)
            probabilities = torch.softmax(outputs, dim=1)
            
        # Generate random confidence scores for demonstration
        confidence_scores = np.random.dirichlet(np.ones(3))
        predicted_class = self.classes[np.argmax(confidence_scores)]
        
        return {
            'prediction': predicted_class,
            'confidence': float(np.max(confidence_scores)),
            'all_scores': {
                class_name: float(score) 
                for class_name, score in zip(self.classes, confidence_scores)
            }
        } 