requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
torch==2.2.1
torchvision==0.17.1
python-dotenv==1.0.0
pydantic==2.5.2
plotly==5.18.0
matplotlib==3.8.2 